fix: Move BacklogLocation enum to own file
