+++ google-cloud-pubsublite/src/main/java/com/google/cloud/pubsublite/BacklogLocation.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.pubsublite;
+
+/** BacklogLocation refers to a location with respect to the message backlog. */
+public enum BacklogLocation {
+  /** BEGINNING refers to the location of the oldest retained message. */
+  BEGINNING,
+  /**
+   * END refers to the location past all currently published messages, skipping the entire message
+   * backlog.
+   */
+  END
+}
